+++ scripts/content.mjs
@@ -1,7 +1,10 @@
 import { existsSync, mkdirSync, writeFileSync } from 'fs'
 import { join } from 'path'
-import { getFilename, getToday, gitNewBranch, openInVSCode } from './common.mjs'
 import { frontmatter } from './frontmatter.mjs'
+import { getFilename } from './getFilename.mjs'
+import { getToday } from './getToday.mjs'
+import { gitNewBranch } from './gitNewBranch.mjs'
+import { openInVSCode } from './openInVsCode.mjs'
 
 const newContent = async () => {
   const { fileName, type, contentFolder } = getFilename()
@@ -9,25 +12,25 @@ const newContent = async () => {
   console.log(`Trying to create a new ${type} content: ${fileName}.mdx`)
 
   const [year, month] = getToday()
-  const filePath = join(contentFolder, year, month, fileName)
+  const folderPath = join(contentFolder, year, month, fileName)
 
-  if (!existsSync(filePath)) {
-    mkdirSync(filePath, { recursive: true })
+  if (!existsSync(folderPath)) {
+    mkdirSync(folderPath, { recursive: true })
   }
 
-  const pathAndFileName = `${filePath}/${fileName}.mdx`
+  const folderAndFilename = `${folderPath}/${fileName}.mdx`
 
-  if (existsSync(pathAndFileName)) {
+  if (existsSync(folderAndFilename)) {
     throw new Error(
       "There's already a file with that name in this month's folder"
     )
   }
 
-  writeFileSync(pathAndFileName, frontmatter[type])
+  writeFileSync(folderAndFilename, frontmatter[type])
   console.log(`${fileName}.mdx created succesfully!`)
 
   await gitNewBranch(type, fileName)
-  openInVSCode(pathAndFileName)
+  openInVSCode(folderAndFilename)
 }
 
 await newContent()
